+++ src/lib/pages/account-details/components/tables/redelegation/index.tsx
@@ -2,7 +2,7 @@ import { TableContainer } from "@chakra-ui/react";
 
 import { useMobile } from "lib/app-provider";
 import { MobileTableContainer } from "lib/components/table";
-import type { Redelegation } from "lib/pages/account-details/data";
+import type { Redelegation } from "lib/pages/account-details/types";
 
 import { RedelegationTableHeader } from "./RedelegationTableHeader";
 import { RedelegationTableMobileCard } from "./RedelegationTableMobileCard";