+++ src/lib/services/contract.ts
@@ -4,8 +4,6 @@ import { z } from "zod";
 import type {
   Addr,
   ContractAddr,
-  Option,
-  PublicInfo,
   ContractInfo,
   ContractMigrationHistory,
 } from "lib/types";
@@ -15,9 +13,9 @@ import {
   zContractHistoryRemark,
   zUtcDate,
 } from "lib/types";
-import { encode, libDecode } from "lib/utils";
+import { encode, libDecode, snakeToCamel } from "lib/utils";
 
-export interface ContractCw2InfoRaw {
+interface ContractCw2InfoRaw {
   data: string;
 }
 
@@ -40,14 +38,6 @@ export interface ContractResponse {
     ibc_port_id: string;
     extension?: string;
   };
-}
-
-interface PublicInfoResponse {
-  slug: string;
-  name: string;
-  address: ContractAddr;
-  description: string;
-  github: string;
 }
 
 export const queryData = async (
@@ -80,25 +70,6 @@ export const queryContractCw2Info = async (
     `${endpoint}/cosmwasm/wasm/v1/contract/${contractAddress}/raw/Y29udHJhY3RfaW5mbw%3D%3D`
   );
   return JSON.parse(libDecode(data.data)) as ContractCw2Info;
-};
-
-export const queryPublicInfo = async (
-  chainName: string | undefined,
-  chainId: string | undefined,
-  contractAddress: ContractAddr
-): Promise<Option<PublicInfo>> => {
-  if (!chainName || !chainId)
-    throw new Error("Invalid chain (queryPublicInfo)");
-  return axios
-    .get<PublicInfoResponse[]>(
-      `https://cosmos-registry.alleslabs.dev/data/${chainName}/${chainId}/contracts.json`
-    )
-    .then(({ data }) => {
-      const publicInfo = data.find((info) => info.address === contractAddress);
-      return publicInfo
-        ? { ...publicInfo, contractAddress: publicInfo.address }
-        : undefined;
-    });
 };
 
 const zContractsResponseItem = z
@@ -205,3 +176,25 @@ export const getMigrationHistoriesByContractAddress = async (
       },
     })
     .then(({ data }) => zMigrationHistoriesResponse.parse(data));
+const zContractTableCounts = z
+  .object({
+    tx: z.number().nullish(),
+    migration: z.number().nullish(),
+    related_proposal: z.number().nullish(),
+  })
+  .transform(snakeToCamel);
+
+export type ContractTableCounts = z.infer<typeof zContractTableCounts>;
+
+export const getContractTableCounts = async (
+  endpoint: string,
+  address: string,
+  isGov: boolean
+): Promise<ContractTableCounts> =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(address)}/table-counts`, {
+      params: {
+        is_gov: isGov,
+      },
+    })
+    .then((res) => zContractTableCounts.parse(res.data));