+++ src/lib/services/move/resourceService.ts
@@ -15,19 +15,19 @@ import { truncate } from "lib/utils";
 
 import { getAccountResources } from "./resource";
 
-export interface AccountResourcesReturn {
+export interface ResourcesByAddressReturn {
   groupedByOwner: ResourceGroupByAccount[];
   groupedByName: ResourceGroup[];
   totalCount: number;
 }
 
-export const useAccountResources = (
+export const useResourcesByAddress = (
   address: MoveAccountAddr
-): UseQueryResult<AccountResourcesReturn> => {
+): UseQueryResult<ResourcesByAddressReturn> => {
   const endpoint = useBaseApiRoute("accounts");
   const { enabled } = useMoveConfig({ shouldRedirect: false });
 
-  const queryFn: QueryFunction<AccountResourcesReturn> = () =>
+  const queryFn: QueryFunction<ResourcesByAddressReturn> = () =>
     getAccountResources(endpoint, address).then((resources) => {
       const groupedByOwner = resources.items.reduce<
         Record<string, ResourceGroupByAccount>
@@ -84,7 +84,7 @@ export const useAccountResources = (
       };
     });
   return useQuery(
-    [CELATONE_QUERY_KEYS.ACCOUNT_RESOURCES, endpoint, address],
+    [CELATONE_QUERY_KEYS.RESOURCES_BY_ADDRESS, endpoint, address],
     queryFn,
     { enabled, refetchOnWindowFocus: false, retry: 1 }
   );