+++ CHANGELOG.md
@@ -39,6 +39,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+### Improvements
+
+- [#697](https://github.com/alleslabs/celatone-frontend/pull/697) api v1 - contract tables (migrations and related proposals)
+- [#696](https://github.com/alleslabs/celatone-frontend/pull/696) api v1 - block details
+- [#678](https://github.com/alleslabs/celatone-frontend/pull/678) api v1 - contract table counts
+
+### Bug fixes
+
+## v1.4.0
+
+### Features
+
 - [#686](https://github.com/alleslabs/celatone-frontend/pull/686) Add amp module list, my published module, and resources/module tab
 - [#664](https://github.com/alleslabs/celatone-frontend/pull/664) Add Vercel speed insights and analytics
 - [#677](https://github.com/alleslabs/celatone-frontend/pull/677) Add minimove-1 and miniwasm-1
@@ -54,8 +66,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
-- [#697](https://github.com/alleslabs/celatone-frontend/pull/697) api v1 - contract tables (migrations and related proposals)
-- [#696](https://github.com/alleslabs/celatone-frontend/pull/696) api v1 - block details
 - [#667](https://github.com/alleslabs/celatone-frontend/pull/667) Adjust for search bar consistency
 - [#692](https://github.com/alleslabs/celatone-frontend/pull/692) Add verify link for mahalo
 - [#675](https://github.com/alleslabs/celatone-frontend/pull/675) Make module links clickable with command/ctrl and add copier for struct names
@@ -82,6 +92,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Bug fixes
 
+- [#698](https://github.com/alleslabs/celatone-frontend/pull/698) Update Neutron gas price
 - [#694](https://github.com/alleslabs/celatone-frontend/pull/694) Reroute cosmwasm pool back to pool detail page
 - [#691](https://github.com/alleslabs/celatone-frontend/pull/691) Fix unsupported token card
 - [#693](https://github.com/alleslabs/celatone-frontend/pull/693) Fix ui styling for contract card and states